Implement basic login sequence
withAuthenticationApp now checks for a username and password before
presenting PostContainer. Checks username and password against local
storage. If new user, creates user and logins. If existing user, tests
for password.

Once logged in, the main App is presented, and there is a props.user on
App that can be used by those components' methods