Add search functionality
I included the ability to search posts based on username. The way the Readme was worded made it sound like we needed to update the state to only contain the search results. I don't really like that idea as it removes all items from the data. This could be solved using localhost storage. I will use the remaining time to explore that and add the ability to clear search and revert back to default page view.