Refactors All Login Logic To Be In Authenticate
I wasn't really sure how to handle this but after looking back at the
authentication component, the login information and all of the user
stuff should probably on that.
watching video from lambda school had the entire app wrapped by
the authentication component. my structure is set up so that the app
contains the authenticated component which then delegates things to get
the login page or the actual posts. I'm wondering which way
is best practice.