Complete MVP for Day 3
I completed the last of the requirements for the third day of this project. The instructions were a bit unclear, and in one case wrong. The login function was not supposed to be in app.js, but in Login.js. Once I got through the difficulty with the instructions, the coding itself was not too bad. Tomorrow, I have time to go through and explore the stretch goals.